welcome window

+++ DLL/Book.cs
@@ -12,6 +12,6 @@ namespace DLL
         public BGenre Genre { get; set; }
         public int YearPublished { get; set; }
         public int Pages { get; set; }
-        public int ISBN { get; set; }
+        public long ISBN { get; set; }
     }
 }
